requirements.txt
opencv-python==4.4.0.46
Pillow==8.0.1
pony==0.7.14
pydub==0.24.1
requests==2.24.0

pip~=22.0.4
setuptools~=49.2.1
freezegun~=1.2.0